Get the test case passing
Small values are working really well now, but there is a massive jump in
errors once we get to the intermediate "no correction" range. I assume
that it would get more accurate as the size becomes larger within that
range (i.e. the corrections will be "most incorrect" at the boundaries
of these conditions). Unfortunately the two hundred thousand odd words
in my dictionary are on the low end of "intermediate" so I cannot be too
sure.